Feat : apply responsive design at bestCafe and bestContent
