+++ src/components/PopularCafeItem/index.js
@@ -1,0 +1,1 @@
+export { default as PopularCafeItem } from './PopularCafeItem';
