+++ src/components/PopularCafes/PopularCafes.jsx
@@ -2,29 +2,21 @@ import {
   Container,
   Content,
   DogCafeList,
-  DogCafeListItem,
-  DogCafeListItemImg,
-  DogCafeListItemTitle,
-  DogCafeListItemRating,
-  DogCafeInfoContainer,
   ErrorContainer,
   More,
+  Title,
 } from './styles';
 import { useGetPolularCafes } from '../../hooks/getPopularCafes';
 import { useNavigate } from 'react-router-dom';
 import { ROUTE } from '../../routes/Routes';
 import { GiPartyPopper } from 'react-icons/gi';
 import { Loading } from '../Loading';
+import { PopularCafeItem } from '../PopularCafeItem';
 
 const PopularCafes = () => {
   const navigate = useNavigate();
 
-  const linkToCafe = (id) => {
-    navigate(`${ROUTE.CAFE_DETAIL_PAGE.link}/${id}`);
-    window.scrollTo({ top: 0, left: 0, behavior: 'smooth' });
-  };
-
-  const linkTop100 = (id) => {
+  const linkTop100 = () => {
     navigate(ROUTE.Popular_Cafe_List_Page.link);
     window.scrollTo({ top: 0, left: 0, behavior: 'smooth' });
   };
@@ -33,66 +25,32 @@ const PopularCafes = () => {
 
   return isLoading ? (
     <Loading />
-  ) : error ? (
-    <Container>
-      <Content>
-        <GiPartyPopper size={'43px'} color="rgba(255, 77, 77, 0.94)" />
-        인기 애견카페
-        <GiPartyPopper
-          size={'43px'}
-          color="rgba(255, 77, 77, 0.94)"
-          style={{ transform: 'scaleX(-1)' }}
-        />
-      </Content>
-      <ErrorContainer>
-        인기 애견카페를 불러오는 도중 에러가 생겼습니다.
-      </ErrorContainer>
-    </Container>
   ) : (
     <Container>
-      <Content>
-        <GiPartyPopper size={'40px'} color="rgba(255, 77, 77, 0.94)" />
-        <span>인기 애견카페</span>
-        <GiPartyPopper
-          size={'40px'}
-          color="rgba(255, 77, 77, 0.94)"
-          style={{ transform: 'scaleX(-1)' }}
-        />
-      </Content>
-      <More
-        onClick={() => {
-          linkTop100();
-        }}
-      >
-        더보기
-      </More>
-      <DogCafeList>
-        {data &&
-          data.cafes.map((content, index) => (
-            <DogCafeListItem
-              key={index}
-              onClick={() => {
-                linkToCafe(content._id);
-              }}
-            >
-              <DogCafeListItemImg
-                src={
-                  content.image.length > 0
-                    ? content.image
-                    : '/imges/default.png'
-                }
-                alt={`cafe ${index}`}
-              />
-              <DogCafeInfoContainer>
-                <DogCafeListItemTitle>{content.name}</DogCafeListItemTitle>
-                <DogCafeListItemRating>
-                  평균 별점
-                  <span>{content.rating}</span>점
-                </DogCafeListItemRating>
-              </DogCafeInfoContainer>
-            </DogCafeListItem>
-          ))}
-      </DogCafeList>
+      <Title>
+        <Content>
+          <GiPartyPopper size={'40px'} color="rgba(255, 77, 77, 0.94)" />
+          <span>인기 애견카페</span>
+          <GiPartyPopper
+            size={'40px'}
+            color="rgba(255, 77, 77, 0.94)"
+            style={{ transform: 'scaleX(-1)' }}
+          />
+        </Content>
+        <More onClick={linkTop100}>더보기</More>
+      </Title>
+      {error ? (
+        <ErrorContainer>
+          인기 애견카페를 불러오는 도중 에러가 생겼습니다.
+        </ErrorContainer>
+      ) : (
+        <DogCafeList>
+          {data &&
+            data.cafes.map((cafe) => (
+              <PopularCafeItem key={cafe._id} cafe={cafe} />
+            ))}
+        </DogCafeList>
+      )}
     </Container>
   );
 };
